use 64bit float

+++ types/types.go
@@ -7,6 +7,6 @@ type BusTipInfo struct {
 
 type BusGPSData struct {
 	Name      string
-	Latitude  float32
-	Longitude float32
+	Latitude  float64
+	Longitude float64
 }
